Resolve review
