+++ identity-did/src/verification/mod.rs
@@ -8,13 +8,14 @@ mod method_query;
 mod method_ref;
 mod method_scope;
 mod method_type;
+mod traits;
 
 pub use self::builder::MethodBuilder;
 pub use self::method::Method;
-pub use self::method::MethodUriType;
-pub use self::method::TryMethod;
 pub use self::method_data::MethodData;
 pub use self::method_query::MethodQuery;
 pub use self::method_ref::MethodRef;
 pub use self::method_scope::MethodScope;
 pub use self::method_type::MethodType;
+pub use self::traits::MethodUriType;
+pub use self::traits::TryMethod;
